sent analysis func
