+++ civiclens/nlp/topics.py
@@ -1,56 +1,12 @@
 from collections import defaultdict
 
 import numpy as np
-import polars as pl
-import torch
 from bertopic import BERTopic
 from langchain_community.vectorstores.utils import maximal_marginal_relevance
 from nltk.tokenize import sent_tokenize
 from sentence_transformers import SentenceTransformer
-from sentence_transformers.util import cos_sim
 
 from ..utils.ml_utils import clean_comments
-
-
-def extract_formletters(
-    docs: list[str],
-    embeddings: torch.tensor,
-    sim_threhold: float = 0.99,
-) -> dict[str, int]:
-    """
-    Extracts from letters from collection of comments.
-
-    Inputs:
-        docs: List of comments
-        embeddings: Embedding matrix for list of comments
-        sim_threhold: Threhold (cosine similarity) for similar comments
-
-    Returns:
-        Dictionary containing text of identified form letters, and number
-        of form comments.
-    """
-    df = pl.DataFrame({"comment": docs})
-    df = df.with_row_index()
-
-    form_comments = {}
-
-    for _ in range(len(docs) // 100):
-        row = df.sample(n=1)
-        idx = row["index"]
-        sim_tensor = cos_sim(embeddings[idx], embeddings)
-
-        _, indices = torch.where(sim_tensor > sim_threhold)
-        sample = df.filter(~df["index"].is_in(indices.tolist()))
-
-        if (len(df) - len(sample)) / len(df) > 0.25:
-            form_comments[row["comment"].item()] = len(df) - len(sample)
-
-        df = sample
-
-        if df.is_empty():
-            break
-
-    return form_comments
 
 
 def mmr_sort(terms: list[str], query_string: str, lam: float) -> list[str]:
